+++ src/utils/data/SubjectData.tsx
@@ -1,0 +1,7 @@
+const SubjectData = [
+  { value: 'english', label: 'English' },
+  { value: 'math', label: 'Math' },
+  { value: 'science', label: 'Science' },
+];
+
+export default SubjectData;
